re-org and fix bug
